fix: benchmarks
